Update Ex8.java
Correção de int para double

+++ Ex8.java
@@ -5,12 +5,12 @@ public class Ex8 {
         Scanner ler = new Scanner(System.in);
 
         System.out.println("informe 2 valores: ");
-        int valorA = ler.nextInt();
-        int valorB = ler.nextInt();
+        double valorA = ler.nextDouble();
+        double valorB = ler.nextDouble();
 
         while(valorA < 0 || valorB == 0){
             System.out.println("Informe um valor maior que zero:");
-            valorB = ler.nextInt();
+            valorB = ler.nextDouble();
         }
         System.out.println("A divisão entre os dois valores é = " + valorA/valorB);
         ler.close();
